+++ src/test/kotlin/org/kalasim/test/TimeTrafoTests.kt
@@ -2,10 +2,7 @@ package org.kalasim.test
 
 import io.kotest.matchers.shouldBe
 import org.junit.Test
-import org.kalasim.Component
-import org.kalasim.Event
-import org.kalasim.OffsetTransform
-import org.kalasim.TickTransform
+import org.kalasim.*
 import java.time.Duration
 import java.time.Instant
 import java.util.concurrent.TimeUnit
@@ -15,50 +12,35 @@ class TimeTrafoTests {
     class TimeTrafoTestEvent(time: Double) : Event(time)
 
     @Test
-    fun `it should project ticks to the real world clock with custon trafo`() = createTestSimulation {
-        val baseTime = Instant.parse("2021-01-24T11:16:00.00Z")
-
-        object : Component() {
-            override fun process() = sequence<Component> {
-                tickTransform = TickTransform { baseTime + Duration.ofSeconds(now.toLong()) }
-
-                addEventListener {
-                    transformTickTime(now) shouldBe transformTickTime(it.time)
-                }
-
-                hold(100)
-                log(TimeTrafoTestEvent(now))
-            }
-        }
-
-        run(1000)
-    }
-
-    @Test
     fun `it should correctly project simulation times with offset-trafo`() = createTestSimulation(true) {
-        val baseTime = Instant.parse("2021-01-24T11:16:00.00Z")
+        val baseTime = Instant.parse("2021-01-24T12:00:00.00Z")
+
+        tickTransform = OffsetTransform(baseTime, TimeUnit.MINUTES)
 
         object : Component() {
-            override fun process() = sequence<Component> {
-                tickTransform = OffsetTransform(baseTime, TimeUnit.MINUTES)
+            override fun process() = sequence {
 
                 addEventListener {
-                    transformTickTime(now) shouldBe transformTickTime(it.time)
-                    println("tick time is ${transformTickTime(now)}")
+                    if(it !is TimeTrafoTestEvent) return@addEventListener
+
+                    println("tick time is ${asWallTime(now)}")
+
+                    asWallTime(now) shouldBe asWallTime(it.time)
+                    asWallTime(now) shouldBe Instant.parse("2021-01-24T13:30:00.00Z")
                 }
 
-                hold(100)
+                hold(asTickDuration(Duration.ofMinutes(90)))
                 log(TimeTrafoTestEvent(now))
             }
         }
 
-        run(1000)
+        run(10000)
+        println()
     }
 
 
     @Test
     fun `it should correctly project real time simulation ticks to the current wall time`() = createTestSimulation {
-
-
+        // implement me
     }
 }